+++ NEWS.md
@@ -1,9 +1,14 @@
 # activAnalyzer (development version)
 * Added scores for the D-PPAC questionnaire.
-* Allowed the possiblity to work with epochs inferior to 60 s (it seems that epochs different from 10/60s do not work).
+* Allowed the possiblity to work with epochs shorter than 60 s. A warning is provided when the set epoch is <10 s to indicate that figures will not be created with such epochs to save time.
 * Updated guide.
 * The user can now set "23:59:59" rather than "23:59:00" to set the upper limit of the period of the day to consider for analyzis.
 * Increased the size of the files that the app can manage.
+* Corrected bug for the message that should appear when a file with incorrect extension is uploaded.
+* Added feedbacks to guide the user for the choice of the values to set to configure the desired epoch for analysis (the ratio between the desired epoch and the current epoch in the file can only be an integer).
+* Replaced for loop by vectorization to detect bouts in the data file with the mark_intensity() function.
+* Removed actigraph.sleepr from the DESCRIPTION file and added the read_agd.R file (from the actigraph.sleepr package) as additionnal file into the activAnalyzer package.
+* The Reset button is now at the top of the app.
 
 # activAnalyzer 0.4.1
 * Updated calculations for PROactive accelerometer scores (previous calculations actually were appropriate when using the means of the days of the measurement, not the medians). Now the app allows to compute scores from both medians and means of the days of the measurement (see references in the documentation for further details).